Fix `stargate#ok_vim` with a pattern which matches only 1 destination
### Summary

When there is only 1 destination, `stargate#ok_vim` with a number argument immediately goes to the destination. But if a
pattern (string) argument is given, `stargate#ok_vim` doesn't work when there is only 1 destination. Patterns which
match multiple destinations are always OK. This pull request fixes this behavior.

### Steps to reproduce the behavior

1. [Shell] start Vim where vim9-stargate is available
2. [Vim] open vim9-stargate's `autoload/stargate.vim`
3. [Vim] execute `:call stargate#ok_vim('#g')`

Expected behavior is that the cursor moves to `#` in `def stargate#galaxy()` (line 57). But actual behavior is that no
destinations are displayed and the cursor doesn't move.